headroom

+++ fastfield_codecs/src/null_index/mod.rs
@@ -2,10 +2,14 @@ pub use dense::{serialize_dense_codec, DenseCodec};
 
 mod dense;
 
-fn get_bit_at(input: u64, n: u32) -> bool {
+
+
+#[inline(always)]
+fn get_bit_at(input: u32, n: u32) -> bool {
     input & (1 << n) != 0
 }
 
-fn set_bit_at(input: &mut u64, n: u64) {
+#[inline(always)]
+fn set_bit_at(input: &mut u32, n: u32) {
     *input |= 1 << n;
 }
